requirements.txt
psycopg[binary]>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
timezonefinder>=6.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

